requirements.txt
google-generativeai
requests
python-dotenv
ffmpeg-python

